+++ Cargo.toml
@@ -17,7 +17,6 @@ members = [
     "sunscreen_compiler_macros",
     "sunscreen_fhe_program",
     "sunscreen_runtime",
-    "sunscreen_zkp_compiler",
     "sunscreen_compiler_common",
     "sunscreen_bulletproofs_backend"
 ]
