+++ sunscreen/Cargo.toml
@@ -24,11 +24,11 @@ bumpalo = "3.8.0"
 log = "0.4.14"
 num = "0.4.0"
 petgraph = "0.6.0"
+sunscreen_compiler_common = { path = "../sunscreen_compiler_common" }
 sunscreen_compiler_macros = { version = "0.7", path = "../sunscreen_compiler_macros" }
 sunscreen_backend = { version = "0.7", path = "../sunscreen_backend" }
 sunscreen_fhe_program = { version = "0.7", path = "../sunscreen_fhe_program" }
 sunscreen_runtime = { version = "0.7", path = "../sunscreen_runtime" }
-sunscreen_zkp_compiler = { path = "../sunscreen_zkp_compiler" }
 seal_fhe = { version = "0.7", path = "../seal_fhe" }
 serde = { version = "1.0.130", features = ["derive"] }
