+++ sunscreen/src/types/zkp/native_field.rs
@@ -1,19 +1,16 @@
-use sunscreen_compiler_common::{Type, TypeName};
+use sunscreen_compiler_macros::TypeName;
 
-use crate::{
-    fe_compiler::ZkpContextOps,
-    types::{AddVar, ProgramNode},
-    with_zkp_ctx,
+use crate::{zkp::{
+    ZkpContextOps, with_zkp_ctx},
+    types::zkp::{AddVar, ProgramNode},
 };
 
 use super::{MulVar, NegVar, NumFieldElements, ZkpType};
 
-use semver::Version;
-
 // Shouldn't need Clone + Copy, but there appears to be a bug in the Rust
 // compiler that prevents ProgramNode from being Copy if we don't.
 // https://github.com/rust-lang/rust/issues/104264
-#[derive(Clone, Copy)]
+#[derive(Clone, Copy, TypeName)]
 /**
  * The native field type in the underlying backend proof system. For
  * example, in Bulletproofs, this is [`Scalar`](https://docs.rs/curve25519-dalek-ng/4.1.1/curve25519_dalek_ng/scalar/struct.Scalar.html).
@@ -53,18 +50,5 @@ impl NegVar for NativeField {
 
             ProgramNode::new(&[o])
         })
-    }
-}
-
-impl TypeName for NativeField {
-    fn type_name() -> Type {
-        let version = env!("CARGO_PKG_VERSION");
-        let version = Version::parse(version).unwrap();
-
-        Type {
-            name: format!("{}::NativeField", module_path!()),
-            version,
-            is_encrypted: false,
-        }
     }
 }
