--- sunscreen_zkp_compiler/src/types/bit.rs
@@ -1,1 +1,0 @@
-pub struct Bit {}
